Debugger: Simplify SymbolTreeLocation struct
Previously each separate object could represent locations from
either the EE or the IOP. We don't need this right now, so I'm
removing this functionality until we do if ever.

Now, the CPU used is passed down from the model and widget objects.
This makes it easier to avoid bugs where the wrong CPU object is used.